+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.2.1
 - `fix` Calculator delete button on pre-Lollipop.
+- `fix` Transaction view on pre-Lollipop.
 - `new` Transaction items are taller now.
 
 ### v0.2.0